pyproject.toml
[project]
name = "yiu_comfy_nodes"
description = "Useful to me. Might be to you."
authors = [
    { name = "leizingyiu" }
]
keywords = ["comfyui", "image-processing"]

version = "0.0.1.20250705"
license = {file = "LICENSE"}
requires-python = ">=3.8"
dependencies = [
    "torch",
    "numpy",
    "Pillow"
]


[project.urls]
Repository = "https://github.com/leizingyiu/yiu_comfy_nodes"

[tool.comfy_custom_node]
PublisherId = "leizingyiu"
DisplayName = "yiu_comfy_nodes"
Icon = ""
entry_point = "yiu_comfy_nodes"
categories = ["Image", "Utility"]
